Update reference for estimateDisp.Rd
Update to Cox Reid C code to allow for large numbers of zero fitted values.

git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/edgeR@90909 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/R_cr_adjust.cpp
@@ -26,7 +26,7 @@ SEXP R_cr_adjust (SEXP w, SEXP x, SEXP nlibs) try {
     double* out_ptr=NUMERIC_POINTER(output);
     try {
 		// Running through each tag.
-       	for (long tag=0; tag<num_tags; ++tag) {
+       	for (int tag=0; tag<num_tags; ++tag) {
 			std::pair<double, bool> x=acr.compute(wptr);
 			if (!x.second) { 
 				std::stringstream errout;
